requirements.txt
action-msgs==1.2.1
action-tutorials-interfaces==0.20.3
action-tutorials-py==0.20.3
actionlib-msgs==4.2.3
ament-cmake-test==1.3.8
ament-copyright==0.12.10
ament-cppcheck==0.12.10
ament-cpplint==0.12.10
ament-flake8==0.12.10
ament-index-python==1.4.0
ament-lint==0.12.10
ament-lint-cmake==0.12.10
ament-package==0.14.0
ament-pep257==0.12.10
ament-uncrustify==0.12.10
ament-xmllint==0.12.10
angles==1.15.0
appdirs==1.4.4
argcomplete==1.8.1
attrs==21.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
boto3==1.34.62
botocore==1.34.62
breezy==3.2.1
Brotli==1.0.9
builtin-interfaces==1.2.1
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
certifi==2020.6.20
chardet==4.0.0
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.1.1
colcon-cmake==0.2.28
colcon-common-extensions==0.3.0
colcon-core==0.15.2
colcon-defaults==0.2.8
colcon-devtools==0.2.5
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.2.15
colcon-output==0.2.13
colcon-package-information==0.3.3
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.8
colcon-recursive-crawl==0.2.3
colcon-ros==0.4.1
colcon-test-result==0.3.8
colcon-zsh==0.5.0
composition-interfaces==1.2.1
configobj==5.0.6
coverage==7.4.3
cryptography==3.4.8
cv-bridge==3.2.1
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defusedxml==0.7.1
demo-nodes-py==0.20.3
Deprecated==1.2.13
diagnostic-msgs==4.2.3
distlib==0.3.8
distro==1.7.0
docutils==0.17.1
domain-coordinator==0.10.0
dulwich==0.20.31
durationpy==0.6
empy==3.3.4
example-interfaces==0.9.3
examples-rclpy-executors==0.15.1
examples-rclpy-minimal-action-client==0.15.1
examples-rclpy-minimal-action-server==0.15.1
examples-rclpy-minimal-client==0.15.1
examples-rclpy-minimal-publisher==0.15.1
examples-rclpy-minimal-service==0.15.1
examples-rclpy-minimal-subscriber==0.15.1
exceptiongroup==1.2.0
fastbencode==0.0.5
fastimport==0.9.14
ffmpeg==1.4
flake8==4.0.1
fonttools==4.29.1
fs==2.4.12
gast==0.5.2
geometry-msgs==4.2.3
gpg==1.16.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
image-geometry==3.2.1
importlib-metadata==4.6.4
iniconfig==1.1.1
interactive-markers==2.3.2
jeepney==0.7.1
jmespath==1.0.1
keyring==23.5.0
kiwisolver==1.3.2
lark==1.1.1
laser-geometry==2.4.0
launch==1.0.5
launch-ros==0.19.7
launch-testing==1.0.5
launch-testing-ros==0.19.7
launch-xml==1.0.5
launch-yaml==1.0.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lifecycle-msgs==1.2.1
logging-demo==0.20.3
lxml==4.8.0
lz4==3.1.3+dfsg
map-msgs==2.1.0
matplotlib==3.5.1
mccabe==0.6.1
mercurial==6.1.1
message-filters==4.3.3
mock==4.0.3
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==0.0.0
natsort==8.4.0
nav-msgs==4.2.3
net-tools==0.1.2
netifaces==0.11.0
nose==1.3.7
notify2==0.3.1
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
osrf-pycommon==2.0.2
packaging==21.3
patiencediff==0.2.1
pbr==5.8.0
pcl-msgs==1.0.0
pendulum-msgs==0.20.3
Pillow==9.0.1
pluggy==1.4.0
ply==3.11
psutil==5.9.0
py==1.10.0
pycairo==1.20.1
pycodestyle==2.8.0
pycryptodomex==3.11.0
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.4.0
PyGithub==1.55
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypylon==3.0.1
PyQt5==5.15.6
PyQt5-sip==12.9.1
pytest==8.1.1
pytest-cov==4.1.0
pytest-repeat==0.9.3
pytest-rerunfailures==13.0
python-dateutil==2.8.1
python-gitlab==2.10.1
python-gnupg==0.4.8
python-qt-binding==1.1.2
pythran==0.10.0
pytz==2022.1
PyYAML==5.4.1
pyzed==1.3.0
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.3
rcl-interfaces==1.2.1
rclpy==3.3.12
rcutils==5.1.5
requests==2.25.1
requests-toolbelt==0.9.1
resource-retriever==3.1.1
rmw-dds-common==1.6.0
roman==3.3
ros2-orb-slam3==1.0.0
ros2action==0.18.9
ros2bag==0.15.9
ros2cli==0.18.9
ros2component==0.18.9
ros2doctor==0.18.9
ros2interface==0.18.9
ros2launch==0.19.7
ros2lifecycle==0.18.9
ros2multicast==0.18.9
ros2node==0.18.9
ros2param==0.18.9
ros2pkg==0.18.9
ros2run==0.18.9
ros2service==0.18.9
ros2topic==0.18.9
rosbag2-interfaces==0.15.9
rosbag2-py==0.15.9
rosdistro-modules==0.9.0
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rospkg-modules==1.5.0
rpyutils==0.2.1
rqt==1.1.7
rqt-action==2.0.1
rqt-bag==1.1.4
rqt-bag-plugins==1.1.4
rqt-console==2.0.3
rqt-graph==1.3.0
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
s3transfer==0.10.0
scipy==1.8.0
SecretStorage==3.3.1
sensor-msgs==4.2.3
sensor-msgs-py==4.2.3
shape-msgs==4.2.3
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.1
sros2==0.10.4
statistics-msgs==1.2.1
std-msgs==4.2.3
std-srvs==4.2.3
stereo-msgs==4.2.3
sympy==1.9
teleop-twist-keyboard==2.3.2
tf2-geometry-msgs==0.25.6
tf2-kdl==0.25.6
tf2-msgs==0.25.6
tf2-py==0.25.6
tf2-ros-py==0.25.6
tf2-tools==0.25.6
toml==0.10.2
tomli==2.0.1
topic-monitor==0.20.3
trajectory-msgs==4.2.3
turtlesim==1.4.2
ufoLib2==0.13.1
unicodedata2==14.0.0
unique-identifier-msgs==2.2.1
urllib3==1.26.5
vcstools==0.1.42
visualization-msgs==4.2.3
wadllib==1.3.6
webencodings==0.5.1
wget==3.2
wrapt==1.13.3
wstool==0.1.18
zipp==1.0.0
